epochs: vendor epochs module from osmosis
  github.com/osmosis-labs/osmosis/v25

Ideally, we'd like to use the epochs module directly from the
upstream. However, we keep running into dependencies issues with
more recent version of Cosmos SDK.

So, we vendor the source code, and update it for our adoption.

We can revisit the upstream later or wait until Cosmos SDK adopts
or offer similar modules.